+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-07-03 20:22:15
-// Commit: test(core): implement widget integration tests for core
+// Updated at: 2026-07-04 11:24:51
+// Commit: feat(core): implement responsive core view component
 
 class coreComponentState {
-  final String commitRef = '2551';
-  final String lastMessage = 'test(core): implement widget integration tests for core';
-  final DateTime timestamp = DateTime.parse('2026-07-03T20:22:15+05:00');
+  final String commitRef = '2565';
+  final String lastMessage = 'feat(core): implement responsive core view component';
+  final DateTime timestamp = DateTime.parse('2026-07-04T11:24:51+05:00');
 }